requirements.txt
openai
flask
dotenv_values